requirements.txt
plyfile
loguru
omegaconf
opencv-python
joblib
scipy
libigl
smplx
trimesh
open3d
lpips
numpy==1.23.1
easydict
aitviewer
loguru
lpips
libigl
plyfile
pydantic
diffusers==0.24
transformers
einops
kornia
accelerate
huggingface_hub==0.24.7
pickle5
triton
controlnet_aux
torchmetrics
torch-metrics
numba
wandb
basicsr
mediapipe